Inline resolveInlineChrome back into EditorToolbarArea
